refactored mapper structure
